fix: avoid errors on IE 11 when upper case exists in base
In Chrome, when there is a subdomain in base that is upper-cased, it will return
an URL that is lower-cased. In IE 11 with this polyfills, we get the error
"Error: URL unable to set base <Upper-Case-URL> due to Error: <Lower-Case-Url>".

This commit will make the polyfill match the behavior in Chrome (and others)
by ensuring that base is always lower cased.